requirements.txt
cachetools==2.0.1
certifi==2018.1.18
chardet==3.0.4
dj-database-url==0.5.0
Django==2.0.3
django-heroku==0.3.1
djangorestframework==3.7.7
google-api-core==1.1.0
google-auth==1.4.1
google-cloud-language==1.0.1
googleapis-common-protos==1.5.3
grpcio==1.10.0
gunicorn==19.7.1
idna==2.6
Markdown==2.6.11
protobuf==3.5.2.post1
psycopg2==2.7.4
pyasn1==0.4.2
pyasn1-modules==0.2.1
python-decouple==3.1
pytz==2018.3
requests==2.18.4
rsa==3.4.2
six==1.11.0
urllib3==1.22
whitenoise==3.3.1
